selenium-python/requirements.txt
selenium==4.*
pytest==8.*
python-dotenv==1.*
browserstack-sdk>=1.0.0
requests>=2.31.0
